Reject OAT file in speed-profile if app image is rejected
Speed profile should always have an app image available. This change
makes ART refuse to load OAT files compiled with the speed-profile
compiler filter if the app image cannot be loaded. This will enable the
compiler to assume the existence of an app image and generate better
code in some cases.

Bug: 38313278
Test: ./test/run-test --host 2231-oat-require-app-image
Change-Id: I2ee6fa7c7dc1d42d1fe1511fc60a7c0542237bfa
Signed-off-by: mydongistiny <jaysonedson@gmail.com>